# public/api-docs.yaml
openapi: 3.0.3
info:
  title: Assessment API - Video Platform
  description: A comprehensive video sharing platform API with authentication, video management, social features, and content discovery
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000/api
    description: Development server
security:
  - JWT: []
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        errors:
          type: array
          items:
            type: string
        pagination:
          type: object
paths:
...
  /login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /videos:
    get:
      tags:
        - Videos
      summary: Get public videos
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: Videos retrieved successfully
...
  /videos/search:
    get:
      tags:
        - Videos
      summary: Search videos
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
...
  /search/videos:
    get:
      tags:
        - Search
      summary: Search videos with filters
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: duration
          in: query
          schema:
            type: string
        - name: uploadDate
          in: query
          schema:
            type: string
        - name: creator
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            default: relevance
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Search results
  /search/hashtags/{hashtag}:
    get:
      tags:
        - Search
      summary: Search videos by hashtag
      parameters:
        - name: hashtag
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Hashtag videos retrieved
  /search/suggestions:
    get:
      tags:
        - Search
      summary: Get search suggestions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Search suggestions
  /search/hashtags/trending:
    get:
      tags:
        - Search
      summary: Get trending hashtags
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Trending hashtags retrieved
...
  /profile:
    put:
      tags:
        - User Profile
      summary: Update profile
      security:
        - JWT: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                firstName:
                  type: string
                bio:
                  type: string
                profilePicture:
                  type: string
      responses:
        '200':
          description: Profile updated successfully
...